requirements.txt
pandas
numpy
scikit-learn
matplotlib
seaborn
nltk
scikit-surprise
streamlit
textblob